Factor out ad-blocker detection into `useShouldShowAds`
... for it to be re-used on other ad components.
